refactor the CLI into separate modules
